AddInternship

+++ src/main/java/tn/esprit/innoxpert/Service/InternshipService.java
@@ -3,15 +3,23 @@ package tn.esprit.innoxpert.Service;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import tn.esprit.innoxpert.Entity.Internship;
+import tn.esprit.innoxpert.DTO.AddInternship;
+import tn.esprit.innoxpert.Entity.*;
+import tn.esprit.innoxpert.Exceptions.NotFoundException;
 import tn.esprit.innoxpert.Repository.InternshipRepository;
+import tn.esprit.innoxpert.Repository.PostRepository;
+import tn.esprit.innoxpert.Repository.UserRepository;
 
 import java.util.List;
+import java.util.Map;
+
 @Service
 @AllArgsConstructor
 public class InternshipService implements InternshipServiceInterface {
     @Autowired
     InternshipRepository internshipRepository;
+    UserRepository userRepository;
+    PostRepository postRepository;
 
     @Override
     public List<Internship> getAllInternships() {
@@ -19,17 +27,67 @@ public class InternshipService implements InternshipServiceInterface {
     }
 
     @Override
-    public Internship getInternshipById(Long internshipId) {
-        return internshipRepository.findById(internshipId).orElse(null);
+    public List<Internship> getInternshipsByCriteria(Long idUser, Long idPost) {
+        if (idUser != null) {
+            return internshipRepository.findAll().stream()
+                    .filter(internship -> internship.getUsers().stream()
+                            .anyMatch(user -> user.getIdUser().equals(idUser)))
+                    .toList();
+        }
+        else if (idPost != null) {
+            return internshipRepository.findAll().stream()
+                    .filter(internship -> internship.getPost() != null && internship.getPost().getId().equals(idPost))
+                    .toList();
+        } else {
+            throw new IllegalArgumentException("Au moins un critère (idUser ou idPost) doit être fourni.");
+        }
     }
 
     @Override
-    public Internship addInternship(Internship b) {
-        return null;
+    public Internship getInternshipById(Long internshipId) {
+        return internshipRepository.findById(internshipId)
+                .orElseThrow(() -> new NotFoundException("Meeting with ID : " + internshipId + " was not found."));
+    }
+
+    @Override
+    public Internship addInternship(AddInternship addInternship) {
+        Long userId = addInternship.getIdUser();
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+
+        if (user.getTypeUser() != TypeUser.Student) {
+            throw new RuntimeException("Only students can apply for an internship.");
+        }
+
+        List<Document> documents = user.getDocuments();
+        if (documents == null || documents.isEmpty()) {
+            throw new RuntimeException("User does not have any documents.");
+        }
+
+        Document cvDocument = documents.stream()
+                .filter(doc -> "CV".equalsIgnoreCase(String.valueOf(doc.getTypeDocument())))
+                .findFirst()
+                .orElseThrow(() -> new RuntimeException("User does not have a CV document."));
+
+        Long idPost = addInternship.getIdPost();
+        Post post = postRepository.findById(idPost)
+                .orElseThrow(() -> new NotFoundException("Post not found"));
+        Internship internship = new Internship();
+        internship.setTitle(post.getTitle());
+        internship.setDescription(post.getContent());
+        internship.setInternshipState(InternshipState.PENDING);
+        internship.setId_document(cvDocument.getId());
+        internship.setPost(post);
+        internship.getUsers().add(user);
+
+        return internshipRepository.save(internship);
     }
 
     @Override
     public void removeInternshipById(Long internshipId) {
+        if (!internshipRepository.existsById(internshipId)) {
+            throw new NotFoundException("Meeting with ID :  " + internshipId + " was not found.");
+        }
         internshipRepository.deleteById(internshipId);
     }
 
@@ -39,12 +97,31 @@ public class InternshipService implements InternshipServiceInterface {
     }
 
     @Override
-    public Internship approveInternship(Internship b) {
+    public Internship approveInternship(Long internshipId) {
         return null;
     }
 
     @Override
-    public Internship approveInternshipById(Long internshipId) {
+    public Internship rejectInternship(Long internshipId) {
         return null;
     }
+
+    @Override
+    public Map<String, Object> getInternshipStatistics() {
+        return null;
+    }
+
+    @Override
+    public Internship affectationTutor(Long internshipId) {
+        return null;
+    }
+
+    @Override
+    public Internship GenerateInternshipcertificate(Long interbshipId) {
+        return null;
+    }
+
 }
+
+
+
